requirements.txt
Flask
google-generativeai
requests
typing-extensions
